% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{catch_errors}
\alias{catch_errors}
\title{Catch the errors of an R script.}
\usage{
catch_errors(script)
}
\arguments{
\item{script}{The R script}
}
\value{
A message
}
\description{
Runs a R script in a new environment and catches errors and warnings.
}
