% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_script.R
\name{copy_chapter}
\alias{copy_chapter}
\title{Copy the R code of Practice R chapter.}
\usage{
copy_chapter()
}
\value{
A tibble
}
\description{
The function get the book script from Github
opens a new script and insert the code in the script.
}
