% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_script.R
\name{copy_template}
\alias{copy_template}
\title{Copy a Rmarkdown template.}
\usage{
copy_template(file)
}
\arguments{
\item{file}{Name of script}
}
\value{
Rmd
}
\description{
The function gets the book template from Github
and opens it as a new rmarkdown.
}
