% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gssm5}
\alias{gssm5}
\title{A Subset of the General Social Survey data, 2016}
\format{
A data frame with 2538 rows and 5 variables.
}
\source{
National Opinion Research Center, http://gss.norc.org.
}
\usage{
gssm5
}
\description{
The Subset of the General Social Survey 2016
See http://gss.norc.org/Get-Documentation for full
documentation of the variables.
}
\details{
\itemize{
\item race. race of respondent.
\item sex. respondent's sex.
\item income. total family income.
\item happy. general happiness.
\item age. Age recoded into years.
}
}
\keyword{datasets}
