% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_packages.R
\name{list_package}
\alias{list_package}
\title{List the Names of R Packages Used in Practice R}
\usage{
list_package(file, insert = TRUE)
}
\arguments{
\item{file}{Name of script file}

\item{insert}{If TRUE the libraries will be insert in an R Script}
}
\value{
A character vector
}
\description{
The function returns a character vector with
package names used in Practice R
}
