% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_packages.R
\name{list_packages}
\alias{list_packages}
\title{List the Names of all Packages Used in Practice R}
\usage{
list_packages()
}
\value{
A character vector
}
\description{
The returns a character vector with
all package names used in Practice R
}
