% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_r.R
\name{operators}
\alias{operators}
\title{Operators}
\usage{
operators(operator)
}
\arguments{
\item{operator}{Name of the operator}
}
\value{
Returns operators via the console
}
\description{
Examine logical and relational operators
}
