% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{packages_PR}
\alias{packages_PR}
\title{List of packages}
\format{
An object of class \code{character} of length 16.
}
\source{
Practice R
}
\usage{
packages_PR
}
\description{
A vector with names of all packages in Practice R.
}
\keyword{datasets}
