% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_r.R
\name{practice}
\alias{practice}
\title{practice}
\usage{
practice(name)
}
\arguments{
\item{name}{name of the tutorial}
}
\value{
loads the tutorial
}
\description{
Practice R - An interactive textbook for
for the social sciences has several learnr tutorials; the
practice function is a convient way to load them
}
