% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_script.R
\name{show_chapters}
\alias{show_chapters}
\title{show_chapters}
\usage{
show_chapters()
}
\value{
A tibble
}
\description{
fetches the book script from Github
opens a new script and insert the code in the script
}
