% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_graphs.R
\name{show_linetypes}
\alias{show_linetypes}
\title{Shows the Line Types in R}
\usage{
show_linetypes()
}
\value{
A ggplot2 object
}
\description{
The show_plot returns a graph that shows
the line types in R
}
