% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_link.R
\name{show_links}
\alias{show_links}
\title{All links from the Practice R Book.}
\usage{
show_links()
}
\value{
Prints data frame
}
\description{
The function copies the URL and
opens the link in the browser.
}
