% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_packages.R
\name{show_packages}
\alias{show_packages}
\title{Returns the Names of R Packages Used in Practice R}
\usage{
show_packages(file, insert = TRUE)
}
\arguments{
\item{file}{Name of script file}

\item{insert}{If TRUE the libraries will be insert in an R Script}
}
\value{
A character vector
}
\description{
The show_packages returns a character vector with
package names used in Practice R
}
