% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_graphs.R
\name{show_plot}
\alias{show_plot}
\title{Shows the linetypes and shapestypes in R}
\usage{
show_plot(x)
}
\arguments{
\item{x}{Plot name}
}
\value{
A ggplot2 object
}
\description{
The show_plot returns a graph that shows
the line types and shape type of base R
}
