% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_graphs.R
\name{show_shapetypes}
\alias{show_shapetypes}
\title{Shows the Shape Types in R}
\usage{
show_shapetypes()
}
\value{
A ggplot2 object
}
\description{
The show_shapetypes function returns a graph that shows
the shape type of base R
}
