% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_r.R
\name{show_tutorial}
\alias{show_tutorial}
\title{Show the Practice R Tutorials}
\usage{
show_tutorial(file)
}
\arguments{
\item{file}{Name of the tutorial}
}
\value{
Returns the tutorial
}
\description{
Practice R comes with several learnr tutorials; the
function is a convenient way to load them
}
