% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_script.R
\name{template}
\alias{template}
\title{Copy a Rmarkdown template.}
\usage{
template(file)
}
\arguments{
\item{file}{Name of script}
}
\value{
Rmd
}
\description{
The function scrapes the book template from Github
and opens the template as a new rmarkdown.
}
