% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployment_raw}
\alias{unemployment_raw}
\title{Unemployment Data Raw}
\format{
An object of class \code{character} of length 1.
}
\source{
Practice R
}
\usage{
unemployment_raw
}
\description{
Raw text data from OECD Labour Force Statistics 2011-2020
}
\keyword{datasets}
