% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_link.R
\name{view_link}
\alias{view_link}
\title{Opens the Browser with a link from a Practice R book.}
\usage{
view_link(url)
}
\arguments{
\item{url}{URL name}
}
\value{
Prints URL
}
\description{
The function copies the URL and
opens the link in the browser.
}
