% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_r.R
\name{practice}
\alias{practice}
\title{Load PracticeR tutorials}
\usage{
practice(x)
}
\arguments{
\item{x}{insert name of the tutorial}
}
\value{
loads the corresponding learnr tutorial
}
\description{
Load PracticeR tutorials
}
