% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_r.R
\name{practice}
\alias{practice}
\title{Load the Practice R Tutorials}
\usage{
practice(name)
}
\arguments{
\item{name}{Name of the tutorial}
}
\value{
Returns the tutorial in the viewer pane
}
\description{
Practice R comes with several learnr tutorials; the
practice function is a convenient way to load them
}
