% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_graphs.R
\name{show_linetypes}
\alias{show_linetypes}
\title{Show Line Types in R}
\usage{
show_linetypes()
}
\value{
A ggplot2 object
}
\description{
The function returns a graph with
the line types in R
}
\examples{
show_linetypes()
}
