% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_link.R
\name{show_link}
\alias{show_link}
\title{Opens the Browser with a link from a Practice R book.}
\usage{
show_link(url, browse = TRUE)
}
\arguments{
\item{url}{URL name}

\item{browse}{Open browser}
}
\value{
Prints URL
}
\description{
The function copies the URL and
opens the link in the browser.
}
