% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_packages.R
\name{show_packages}
\alias{show_packages}
\title{Returns Packages used in Practice R}
\usage{
show_packages(file)
}
\arguments{
\item{file}{Name of script}
}
\value{
A character vector
}
\description{
The show_packages returns a character vector with
package names used in Practice R
}
