% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_packages.R
\name{show_packages_all}
\alias{show_packages_all}
\title{Returns All Packages used in Practice R}
\usage{
show_packages_all()
}
\value{
A character vector
}
\description{
The show_packages returns a character vector with
all package names used in Practice R
}
