% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_script.R
\name{show_script}
\alias{show_script}
\title{show_script returs book scripts}
\usage{
show_script(file)
}
\arguments{
\item{file}{insert name of script}
}
\value{
loads the corresponding learnr tutorial
}
\description{
show_script returs book scripts
}
