% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_script.R
\name{show_script}
\alias{show_script}
\title{Shows an R Script of the Practice R Book}
\usage{
show_script(file)
}
\arguments{
\item{file}{Name of script}
}
\value{
R script
}
\description{
The function scrapes the book script from GitHub,
opens a new script, and inserts the code in RStudio.
}
