% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_shapetypes}
\alias{show_shapetypes}
\title{show_shapetypes Shows shape types from base R}
\usage{
show_shapetypes()
}
\value{
loads the corresponding learnr tutorial
}
\description{
show_shapetypes Shows shape types from base R
}
