% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{transformer}
\alias{transformer}
\title{transformer}
\usage{
transformer(x, method = NULL)
}
\arguments{
\item{x}{Variable to transform}

\item{method}{Transformation method}
}
\value{
Graph
}
\description{
Transforms a numerical variable into
}
