% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{transformer}
\alias{transformer}
\title{Transform a Numerical Input.}
\usage{
transformer(x, method = NULL)
}
\arguments{
\item{x}{Numerical variable}

\item{method}{Transformation method}
}
\value{
Graph
}
\description{
The function transforms a numerical variable according to the
ladder of powers and displays them with histograms.
}
