% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployment}
\alias{unemployment}
\title{Unemployment Data}
\format{
A data frame with 38 rows and eleven variables.
}
\source{
Practice R
}
\usage{
unemployment
}
\description{
Example data from OECD Labour Force Statistics 2011-2020
}
\details{
\itemize{
\item Country name
\item 2011
\item 2012
\item 2013
\item 2014
\item 2015
\item 2016
\item 2017
\item 2018
\item 2019
\item 2020
}
}
\keyword{datasets}
