% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployment_raw}
\alias{unemployment_raw}
\title{List of all loaded packages}
\format{
A data frame with 16 rows and one variables.
}
\source{
Practice R
}
\usage{
unemployment_raw
}
\description{
A vector with all loaded packages in Practice R.
}
\keyword{datasets}
